Add explicit AVX-512 vectorized popcount kernel
Adds a 4x manually unrolled AVX-512 vectorized popcount kernel.
This is essentially a transplant of the unrolled popcount Highway
kernel in the WIP `vekterli/highway-experiments` branch, but with
all higher-level abstractions replaced with their AVX-512 intrinsic
equivalents, and with the unrolling factor lowered from 8 to 4 to
reduce the tedium involved.

This depends on both `VPOPCNTDQ` and `BMI2` instructions, so it
can't be used below `AVX3_DL` levels. One of the main speedup
sources is the explicit use of parallel, shared-nothing accumulators
which avoids data dependency stalls and lets a deep CPU pipeline
do its thing. Empirically, the compiler auto-vectorizer has a hard
time doing these sorts of stunts even though it may otherwise be
able to vectorize a loop.